requirements.txt
aiohttp==3.11.18
aiogram==3.20.0
telethon==1.34.0
python-decouple==3.8
rarfile==4.1
patool==1.12
tqdm== 4.67.1
jsonschema==4.23.0
pysocks>=1.7.1
python-dotenv==1.1.0
